test(grasql): move test resolver out of the main module
