+++ test/test_helper.exs
@@ -1,1 +1,3 @@
 ExUnit.start()
+# Add test/support directory to the code path for tests
+Code.require_file("support/simple_resolver.ex", __DIR__)
